Rename to sc2ts
Closes #20
