+++ sc2ts/__main__.py
