+++ sc2ts/constants.py
