--- tests/data/usher-metadata-date-test.tsv
@@ -1,44 +1,0 @@
-strain	genbank_accession	date	country	host	completeness	length	Nextstrain_clade	pangolin_lineage	Nextstrain_clade_usher	pango_lineage_usher
-100032|LR824129.1|2020-03-10	LR824129.1	2020-03-10	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100033|LR824076.1|2020-03-10	LR824076.1	2020-03-10	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100034|LR824065.1|2020-03-11	LR824065.1	2020-03	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100035|LR824069.1|2020-03-11	LR824069.1	2020-03-11	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1.142
-100036|LR824079.1|2020-03-11	LR824079.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100037|LR824074.1|2020-03-11	LR824074.1	2020-03	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100039|LR824077.1|2020-03-11	LR824077.1	2020-03-11	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100040|LR824067.1|2020-03-11	LR824067.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100041|LR824075.1|2020-03-11	LR824075.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100042|LR824073.1|2020-03-11	LR824073.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100043|LR824066.1|2020-03-11	LR824066.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100045|LR824068.1|2020-03-11	LR824068.1	2020-03-11	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100047|LR824072.1|2020-03-11	LR824072.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100048|LR824071.1|2020-03-11	LR824071.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100049|LR824078.1|2020-03-11	LR824078.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100050|LR824081.1|2020-03-11	LR824081.1	2020-03-11	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1.367
-100051|LR824082.1|2020-03-11	LR824082.1	2020-03-11	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100052|LR824080.1|2020-03-12	LR824080.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100053|LR824083.1|2020-03-12	LR824083.1	2020-03-12	Switzerland	Homo sapiens		29903	20C	B.1	20C	B.1
-100054|LR824087.1|2020-03-12	LR824087.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100055|LR824093.1|2020-03-12	LR824093.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100056|LR824089.1|2020-03-12	LR824089.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100057|LR824097.1|2020-03-12	LR824097.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100058|LR824088.1|2020-03-12	LR824088.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100059|LR824090.1|2020-03-12	LR824090.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1.142
-100060|LR824094.1|2020-03-12	LR824094.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100061|LR824096.1|2020-03-12	LR824096.1	2020-03-12	Switzerland	Homo sapiens		29903	19A	B	19A	B.33
-100062|LR824095.1|2020-03-12	LR824095.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1.372	20B	B.1.1.372
-100063|LR824098.1|2020-03-12	LR824098.1	2020-03-12	Switzerland	Homo sapiens		29903	20C	B.1	20C	B.1
-100064|LR824085.1|2020-03-12	LR824085.1	2020-03-12	Switzerland	Homo sapiens		29903	20C	B.1	20C	B.1
-100065|LR824084.1|2020-03-12	LR824084.1	2020	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1.142
-100066|LR824092.1|2020-03-12	LR824092.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1.10	20B	B.1.1.10
-100067|LR824086.1|2020-03-12	LR824086.1	2020-03-12	Switzerland	Homo sapiens		29903	20C	B.1	20C	B.1
-100068|LR824091.1|2020-03-12	LR824091.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1.142	20B	B.1.1.142
-100069|LR824101.1|2020-03-12	LR824101.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100070|LR824100.1|2020-03-12	LR824100.1	?	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100071|LR824103.1|2020-03-12	LR824103.1	?	Switzerland	Homo sapiens		29903	19A	B	19A	B.33
-100072|LR824102.1|2020-03-12	LR824102.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100073|LR824099.1|2020-03-12	LR824099.1	2020-03-12	Switzerland	Homo sapiens		29903	20B	B.1.1	20B	B.1.1
-100074|LR824104.1|2020-03-12	LR824104.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100075|LR824223.1|2020-03-12	LR824223.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100076|LR824229.1|2020-03-12	LR824229.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
-100077|LR824227.1|2020-03-12	LR824227.1	2020-03-12	Switzerland	Homo sapiens		29903	20A	B.1	20A	B.1
